finish implementing test data integration for images
