+++ integration-tests/plugin-fastify/gatsby-node.ts
@@ -1,6 +1,7 @@
-const path = require("path");
+import path from "path";
+import type { GatsbyNode } from "gatsby";
 
-exports.createPages = async (gatsbyUtilities) => {
+export const createPages: GatsbyNode["createPages"] = async (gatsbyUtilities) => {
   const {
     actions: { createRedirect, createPage },
     graphql,
@@ -32,10 +33,11 @@ exports.createPages = async (gatsbyUtilities) => {
     component: fakerPostArchive,
     defer: true,
     context: {
+      //@ts-ignore
       posts: result.data.allNameData.nodes,
     },
   });
-
+  //@ts-ignore
   result.data.allNameData.nodes.forEach((node) => {
     createPage({
       path: `/faker/${node.lorem.slug}`,
@@ -85,4 +87,41 @@ exports.createPages = async (gatsbyUtilities) => {
     toPath: "http://example.com/*",
     statusCode: 200,
   });
+};
+
+export const createSchemaCustomization: GatsbyNode["createSchemaCustomization"] = ({
+  actions,
+  schema,
+  reporter,
+}) => {
+  reporter.info("image-cdn-test plugin schemaCustomization");
+  actions.createTypes(
+    schema.buildObjectType({
+      name: `TestImage`,
+      fields: {
+        title: "String",
+      },
+      interfaces: [`Node`, `RemoteFile`],
+    })
+  );
+};
+
+export const sourceNodes: GatsbyNode["sourceNodes"] = async ({ actions, reporter }) => {
+  reporter.info("image-cdn-test plugin sourceNodes");
+
+  const testImage = {
+    id: `test-image`,
+    url: "https://images.unsplash.com/photo-1650247452475-b5866374545d?ixlib=rb-1.2.1&q=80&fm=jpg&crop=entropy&cs=tinysrgb",
+    mimeType: "image/jpeg",
+    title: "test image",
+    filename: "indonesia.jpg",
+    width: 2666,
+    height: 3996,
+    internal: {
+      type: `TestImage`,
+      contentDigest: `test-image`,
+    },
+  };
+
+  actions.createNode(testImage);
 };
