--- integration-tests/plugin-fastify/plugins/gatsby-test-image-cdn/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": [["babel-preset-gatsby-package"]]
-}
