--- integration-tests/plugin-fastify/plugins/gatsby-test-image-cdn/.gitignore
@@ -1,8 +1,0 @@
-# Yarn 3 cache
-.yarn/*
-!.yarn/cache
-!.yarn/patches
-!.yarn/plugins
-!.yarn/releases
-!.yarn/sdks
-!.yarn/versions
