--- integration-tests/plugin-fastify/plugins/gatsby-test-image-cdn/README.md
@@ -1,1 +1,0 @@
-# gatsby-test-image-cdn
