--- integration-tests/plugin-fastify/plugins/gatsby-test-image-cdn/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "gatsby-test-image-cdn",
-  "packageManager": "yarn@3.2.0",
-  "scripts": {
-    "build": "babel src --out-dir . --ignore \"**/__tests__,**/*.d.ts\" --extensions \".ts,.js\""
-  },
-  "dependencies": {
-    "@babel/cli": "^7.17.6",
-    "@babel/core": "^7.17.9",
-    "babel-preset-gatsby-package": "^2.12.1",
-    "gatsby-plugin-utils": "^3.6.1"
-  }
-}
